adding in the extra sparkles on all fingertips + tried adding scale to change size of stars, but it didnt work...
I think maybe because its scaling to the corner of the page??? but it completely skewed the alignment of the stars on the fingertips. I could try setting the scaling to the centre of the hand? or scaling when its drawn? but I might just make a smaller star that I can call in to be drawn, because it also could be a variation in colour and size! I think that would be much more visually interesting